Use int type for constants filterExtent_ and filterWidth_
This fixes following warning:

  warning: comparison of integer expressions of different signedness: ‘int’ and ‘const size_t’ {aka ‘const long unsigned int’} [-Wsign-compare]

First, I've tried changing the types of `i` and `j` from `int` to
`size_t`, but that resulted in following warning on expressions like
`p0.y + j - filterWidth_`:

  warning: conversion from ‘size_t’ {aka ‘long unsigned int’} to ‘lass::prim::Vector2D<double>::TValue’ {aka ‘double’} may change value [-Wconversion]

+++ src/output/filter_mitchell.h
@@ -59,8 +59,8 @@ private:
 
 	typedef PerThreadBuffer<OutputSample> TOutputBuffer;
 
-	static constexpr size_t filterWidth_ = 2;
-	static constexpr size_t filterExtent_ = 2 * filterWidth_ + 1;
+	static constexpr int filterWidth_ = 2;
+	static constexpr int filterExtent_ = 2 * filterWidth_ + 1;
 	static constexpr size_t filterFootprint_ = filterExtent_ * filterExtent_;
 	static constexpr size_t bufferSize_ = filterFootprint_ * 16;
